+++ src/FeatureMatcher.cpp
@@ -27,15 +27,18 @@ using std::shared_ptr;
 using std::vector;
 
 FeatureMatcher::FeatureMatcher(float thDistMatchMax, bool bUseLoweRatioTest,
-                               float thRatioTest) :
+                               float thRatioTest, float thAngMatchMax,
+                               int thDistDescMax) :
     mThDistMatchMax(thDistMatchMax), mbUseLoweRatioTest(bUseLoweRatioTest),
-    mThRatioTest(thRatioTest)
+    mThRatioTest(thRatioTest), mThAngMatchMax(thAngMatchMax),
+    mThDistDescMax(thDistDescMax)
 {
     // initialize feature matcher
     if (mbUseLoweRatioTest) {
         // FLANN-based matcher & use Lowe's ratio test
-        mpFeatMatcher = make_shared<cv::FlannBasedMatcher>(
-            cv::makePtr<cv::flann::LshIndexParams>(12, 20, 2));
+        //mpFeatMatcher = make_shared<cv::FlannBasedMatcher>(
+        //    cv::makePtr<cv::flann::LshIndexParams>(12, 20, 2));
+        mpFeatMatcher = cv::BFMatcher::create(cv::NORM_HAMMING, false);
     } else {
         // use symmetric test
         mpFeatMatcher = cv::BFMatcher::create(cv::NORM_HAMMING, true);
@@ -53,6 +56,47 @@ std::vector<cv::DMatch> FeatureMatcher::match2Dto2D(
                             vvMatches21,
                             nBestMatches); // number of best matches
     return filterMatchResult(vvMatches21, pF2, pF1);
+}
+
+std::vector<cv::DMatch> FeatureMatcher::match2Dto2DCustom(
+    const std::shared_ptr<FrameBase>& pF2,
+    const std::shared_ptr<FrameBase>& pF1) const
+{
+    vector<cv::DMatch> vMatches21;
+    vector<cv::KeyPoint> vKpts2 = pF2->keypoints();
+    vMatches21.reserve(vKpts2.size());
+    // traverse each keypoint in frame 2 for its match in frame 1
+    for (unsigned i = 0; i < vKpts2.size(); ++i) {
+        const cv::KeyPoint& kpt2 = vKpts2[i];
+        float angle = kpt2.angle;
+        Mat desc2 = pF2->descriptor(i);
+        vector<int> vKptIndices = pF1->featuresInRange(
+            Mat(kpt2.pt), angle, mThDistMatchMax, mThAngMatchMax);
+        // traverse each keypoint index and get best match
+        int nBestIdx1 = -1; // for frame 1
+        int nBestDist = 256;
+        int nBestDist2nd = 256;
+        for (const int& kptIdx : vKptIndices) {
+            Mat desc1 = pF1->descriptor(kptIdx);
+            int nDist = hammingDistance(desc2, desc1);
+            if (nDist < nBestDist) {
+                nBestDist2nd = nBestDist;
+                nBestDist = nDist;
+                nBestIdx1 = kptIdx;
+            }
+        }
+        if (nBestIdx1 < 0 || nBestDist > mThDistDescMax) {
+            continue;
+        }
+        // Lowe's ratio test
+        if (static_cast<float>(nBestDist) >= mThRatioTest * nBestDist2nd) {
+            continue;
+        }
+        // construct cv::DMatch object
+        cv::DMatch match21(i, nBestIdx1, static_cast<float>(nBestDist));
+        vMatches21.push_back(match21);
+    }
+    return vMatches21;
 }
 
 std::vector<cv::DMatch> FeatureMatcher::match2Dto3D(
@@ -125,6 +169,116 @@ std::vector<cv::DMatch> FeatureMatcher::match2Dto3D(
         }
     }
     return vMatches21;    
+}
+
+std::vector<cv::DMatch> FeatureMatcher::match2Dto3DCustom(
+    const std::shared_ptr<FrameBase>& pF2,
+    const std::shared_ptr<FrameBase>& pF1,
+    bool bBindMPts) const
+{
+    vector<cv::DMatch> vMatches21;
+    map<int, shared_ptr<MapPoint>> mpMPts1 = pF1->getMPtsMap();
+    vMatches21.reserve(mpMPts1.size());
+    // traverse each 3D point for its 2D match
+    for (const auto& pair : mpMPts1) {
+        const shared_ptr<MapPoint>& pMPt = pair.second;
+        Mat x = pF2->coordWorld2Img(pMPt->X3D());
+        // skip out-of-border reprojected points
+        if (!Utility::is2DPtInBorder(x)) {
+            continue;
+        }
+        float angle = pMPt->angle();
+        Mat desc1 = pMPt->descriptor();
+        vector<int> vKptIndices = pF2->featuresInRange(
+            x, angle, mThDistMatchMax, mThAngMatchMax);
+        // traverse each keypoint index and get best match
+        int nBestIdx2 = -1; // for frame 2
+        int nBestDist = 256;
+        int nBestDist2nd = 256;
+        for (const int& kptIdx : vKptIndices) {
+            Mat desc2 = pF2->descriptor(kptIdx);
+            int nDist = hammingDistance(desc2, desc1);
+            if (nDist < nBestDist) {
+                nBestDist2nd = nBestDist;
+                nBestDist = nDist;
+                nBestIdx2 = kptIdx;
+            }
+        }
+        if (nBestIdx2 < 0 || nBestDist > mThDistDescMax) {
+            continue;
+        }
+        // Lowe's ratio test
+        if (static_cast<float>(nBestDist) >= mThRatioTest * nBestDist2nd) {
+            continue;
+        }
+        // construct cv::DMatch object
+        int nIdx1 = pair.first; // frame 1
+        cv::DMatch match21(nBestIdx2, nIdx1, static_cast<float>(nBestDist));
+        vMatches21.push_back(match21);
+    }
+    // update map point data on current frame (pF2)
+    if (bBindMPts) {
+        for (const auto& match21 : vMatches21) {
+            shared_ptr<MapPoint> pMPt = pF1->mappoint(match21.trainIdx);
+            pF2->bindMPt(pMPt, match21.queryIdx);
+        }
+    }    
+    return vMatches21;
+}
+
+std::vector<cv::DMatch> FeatureMatcher::match2Dto3DCustom(
+    const std::shared_ptr<FrameBase>& pF2,
+    const std::vector<std::shared_ptr<MapPoint>>& vpMPts,
+    bool bBindMPts) const
+{
+    vector<cv::DMatch> vMatches21;
+    int nMPts = vpMPts.size();
+    vMatches21.reserve(nMPts);
+    // traverse each 3D point for its 2D match
+    for (int i = 0; i < nMPts; ++i) {
+        const auto& pMPt = vpMPts[i];
+        Mat x = pF2->coordWorld2Img(pMPt->X3D());
+        // skip out-of-border reprojected points
+        if (!Utility::is2DPtInBorder(x)) {
+            continue;
+        }
+        float angle = pMPt->angle();
+        Mat desc1 = pMPt->descriptor();
+        vector<int> vKptIndices = pF2->featuresInRange(
+            x, angle, mThDistMatchMax, mThAngMatchMax);
+        // traverse each keypoint index and get best match
+        int nBestIdx2 = -1; // for frame 2
+        int nBestDist = 256;
+        int nBestDist2nd = 256;
+        for (const int& kptIdx : vKptIndices) {
+            Mat desc2 = pF2->descriptor(kptIdx);
+            int nDist = hammingDistance(desc2, desc1);
+            if (nDist < nBestDist) {
+                nBestDist2nd = nBestDist;
+                nBestDist = nDist;
+                nBestIdx2 = kptIdx;
+            }
+        }
+        if (nBestIdx2 < 0 || nBestDist > mThDistDescMax) {
+            continue;
+        }
+        // Lowe's ratio test
+        if (static_cast<float>(nBestDist) >= mThRatioTest * nBestDist2nd) {
+            continue;
+        }
+        // construct cv::DMatch object
+        int nIdx1 = i; // frame 1
+        cv::DMatch match21(nBestIdx2, nIdx1, static_cast<float>(nBestDist));
+        vMatches21.push_back(match21);
+    }
+    // update map point data on current frame (pF2)
+    if (bBindMPts) {
+        for (const auto& match21 : vMatches21) {
+            shared_ptr<MapPoint> pMPt = vpMPts[match21.trainIdx];
+            pF2->bindMPt(pMPt, match21.queryIdx);
+        }
+    }
+    return vMatches21;
 }
 
 std::vector<cv::DMatch> FeatureMatcher::filterMatchResult(
@@ -253,6 +407,22 @@ cv::Mat FeatureMatcher::getMatchMask2Dto3D(const std::vector<int>& vIdxKpts1,
     mask = mask.t();
     assert(mask.rows == nKpts2 && mask.cols == nKpts1);
     return mask;
+}
+
+int FeatureMatcher::hammingDistance(const cv::Mat& a, const cv::Mat& b) const
+{
+    // Bit set count operation from
+    const int* pa = a.ptr<int32_t>();
+    const int* pb = b.ptr<int32_t>();
+    int nDist = 0;
+    for(int i = 0; i < 8; i++, pa++, pb++)
+    {
+        unsigned  int v = *pa ^ *pb;
+        v = v - ((v >> 1) & 0x55555555);
+        v = (v & 0x33333333) + ((v >> 2) & 0x33333333);
+        nDist += (((v + (v >> 4)) & 0xF0F0F0F) * 0x1010101) >> 24;
+    }
+    return nDist;
 }
 
 } // namespace SLAM_demo