+++ src/FeatureMatcher.hpp
@@ -8,7 +8,7 @@
 #ifndef FEATURE_MATCHER_HPP
 #define FEATURE_MATCHER_HPP
 
-#include <map> // temp
+//#include <map> // temp
 #include <memory>
 #include <vector>
 
@@ -32,9 +32,14 @@ public: // public members
      * @param[in] bUseLoweRatioTest Whether to use Lowe's ratio test.
      * @param[in] thRatioTest       Max threshold of the ratio of smallest 
      *                              feature distance to that of 2nd smallest.
+     * @param[in] thAngMatchMax     Max angle difference between 2 matched 
+     *                              points (in degrees).
+     * @param[in] thDistDescMax     Max feature distance (Hamming) between 2
+     *                              descriptors.
      */
     FeatureMatcher(float thDistMatchMax, bool bUseLoweRatioTest = true,
-                   float thRatioTest = 0.7f);
+                   float thRatioTest = 0.7f, float thAngMatchMax = 15.0f,
+                   int thDistDescMax = 15);
     /** 
      * @brief Match features for 2D-to-2D case between 2 frames.
      * @return A vector of matching keypoints of cv::DMatch type.
@@ -43,7 +48,19 @@ public: // public members
         const std::shared_ptr<FrameBase>& pF2,
         const std::shared_ptr<FrameBase>& pF1) const;
     /** 
-     * @brief Match features for 2D-to-3D case between 2 freames.
+     * @brief Match features for 2D-to-2D case between 2 frames
+     *        (custom implementation).
+     *
+     * The custom implementation will try to get a minimum set of features
+     * for each feature to be matched based on pixel distance and orientation.
+     *
+     * @return A vector of matching keypoints of cv::DMatch type.
+     */
+    std::vector<cv::DMatch> match2Dto2DCustom(
+        const std::shared_ptr<FrameBase>& pF2,
+        const std::shared_ptr<FrameBase>& pF1) const;
+    /** 
+     * @brief Match features for 2D-to-3D case between 2 frames.
      * @param[in] pF2       Pointer to frame 2 which has features for 2D points.
      * @param[in] pF1       Pointer to frame 1 which has features for 3D points.
      * @param[in] bBindMPts Whether to bind newly matched map points to @p pF2.
@@ -56,7 +73,7 @@ public: // public members
         const std::shared_ptr<FrameBase>& pF1,
         bool bBindMPts = true) const;
     /** 
-     * @brief Match features for 2D-to-3D case between 2 freames.
+     * @brief Match features for 2D-to-3D case between 2 frames.
      * @param[in] pF2    Pointer to frame 2 which has features for 2D points.
      * @param[in] vpMPts A std::vector of pointers to 3D map points.
      * @param[in] bBindMPts Whether to bind newly matched map points to @p pF2.
@@ -68,6 +85,42 @@ public: // public members
         const std::shared_ptr<FrameBase>& pF2,
         const std::vector<std::shared_ptr<MapPoint>>& vpMPts,
         bool bBindMPts = true) const;
+    /** 
+     * @brief Match features for 2D-to-3D case between 2 frames 
+     *        (custom implementation).
+     *
+     * The custom implementation will try to get a minimum set of features
+     * for each feature to be matched based on pixel distance and orientation.
+     *
+     * @param[in] pF2       Pointer to frame 2 which has features for 2D points.
+     * @param[in] pF1       Pointer to frame 1 which has features for 3D points.
+     * @param[in] bBindMPts Whether to bind newly matched map points to @p pF2.
+     * @return A vector of matching keypoints of cv::DMatch type, where
+     *         keypoints in frame 2 are in the querying set, and those
+     *         in frame 1 are in the training set.
+     */
+    std::vector<cv::DMatch> match2Dto3DCustom(
+        const std::shared_ptr<FrameBase>& pF2,
+        const std::shared_ptr<FrameBase>& pF1,
+        bool bBindMPts = true) const;
+    /** 
+     * @brief Match features for 2D-to-3D case between 2 frames
+     *        (custom implementation).
+     *
+     * The custom implementation will try to get a minimum set of features
+     * for each feature to be matched based on pixel distance and orientation.
+     *
+     * @param[in] pF2    Pointer to frame 2 which has features for 2D points.
+     * @param[in] vpMPts A std::vector of pointers to 3D map points.
+     * @param[in] bBindMPts Whether to bind newly matched map points to @p pF2.
+     * @return A vector of matching keypoints of cv::DMatch type, where
+     *         keypoints in frame 2 are in the querying set, and those
+     *         in @p vpMPts are in the training set.
+     */
+    std::vector<cv::DMatch> match2Dto3DCustom(
+        const std::shared_ptr<FrameBase>& pF2,
+        const std::vector<std::shared_ptr<MapPoint>>& vpMPts,
+        bool bBindMPts = true) const;
 private: // private data
     /// Feature matcher.
     std::shared_ptr<cv::DescriptorMatcher> mpFeatMatcher;
@@ -75,6 +128,8 @@ private: // private data
     bool mbUseLoweRatioTest; ///< Whether to use Lowe's ratio test.
     /// Max ratio threshold of smallest feature distance to that of 2nd smallest.
     float mThRatioTest;
+    float mThAngMatchMax; ///< Max angle difference between 2 matched keypoints.
+    int mThDistDescMax; ///< Max descriptor distance between 2 descriptors.
 private: // private members
     /**
      * @brief Filter valid feature matching result.
@@ -122,6 +177,15 @@ private: // private members
      */
     cv::Mat getMatchMask2Dto3D(const std::vector<int>& vIdxKpts1,
                                int nKpts2, int nKpts1) const;
+    /**
+     * @brief Compute Hamming distance between 2 256-bit descriptors.
+     * @param[in] a 1st feature descriptor.
+     * @param[in] b 2nd feature descriptor.
+     * @return Hamming distance between descriptor @p a and @p b.
+     * @see http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel 
+     *      for the bit set count operation implemented in this function.
+     */
+    int hammingDistance(const cv::Mat& a, const cv::Mat& b) const;
 };
 
 } // namespace SLAM_demo