+++ src/Map.cpp
@@ -24,7 +24,7 @@ using std::set;
 using std::shared_ptr;
 using std::vector;
 
-const float Map::TH_MIN_RATIO_TRACKED_TO_VISIBLE = 0.3f;
+const float Map::TH_MIN_RATIO_TRACKED_TO_VISIBLE = 0.6f;
 const unsigned Map::TH_MAX_NUM_FRMS_LAST_SEEN = 10000;
 
 void Map::addMPt(const std::shared_ptr<MapPoint>& pMPt)
@@ -79,8 +79,10 @@ void Map::updateKFData(const std::shared_ptr<KeyFrame>& pKF, int cnt)
     auto it = mmpKFs.find(pKF);
     if (it == mmpKFs.end()) {
         // insert new frame pointer
-        assert(cnt == 1);
-        mmpKFs.insert({pKF, cnt});
+        //assert(cnt == 1);
+        if (cnt == 1) {
+            mmpKFs.insert({pKF, cnt});
+        }
     } else {
         assert(it->second > 0);
         it->second += cnt;